Section page style changes (mostly /about)
Updated the `list.html` layout to not include post_meta. I thought it
was confusing -- is this a post, or is it a section? It also pushed the
actual list of posts too far down the page.

Merges the "Organization" and "technology-credits" posts under /about
back into the index page. They were really short and didn't make sense
on their own. I only split them out to try to bring the list of posts
back up above the fold -- but getting rid of post_meta on the section
page did that for me.

Adds a border, padding, & margin to the bottom of the index-content div.