conway's law post
the way to fix (i think, i've never seen someone do this) it is by
having two or three sre leads and then one or more sre engineers on each
product team. all product team leads are given an additional goal which
is to measure the reliability of their products, setup achievable SLOs,
and then improve on those SLOs over time. the _product engineering lead_
is responsible for hitting the product reliability goal and they have
the sre resources to do it.

the sre leads (with help from the embedded sre engineers, like 20-30%
time) develop/maintain/run the shared o11y infrastructure/tooling, work
on org-level o11y, and mentor the embedded sre engineers.

the product team engineers & their sre participate in a very relaxed (i
have a lot to say about this another time) on-call rotation as peers
with the goal of never getting paged. sre leads backstop them all in a
rotation.

this makes reliability the problem of the people building the product.
they're incentivized and empowered to take care of it.